requirement.txt
Flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0
aiohttp==3.9.1
requests==2.31.0
PyJWT==2.8.0
geopy==2.4.1
redis==5.0.1
psycopg2-binary==2.9.9
prometheus-client==0.19.0
bcrypt==4.1.2
python-dotenv==1.0.0